+++ src/DataSGen.h
@@ -158,6 +158,7 @@ class DataSGen: public DataGeneric
     int get_bus(int sgen_id) {return _get_bus(sgen_id, status_, bus_id_);}
     void change_p(int sgen_id, real_type new_p, bool & need_reset);
     void change_q(int sgen_id, real_type new_q, bool & need_reset);
+    virtual void reconnect_connected_buses(std::vector<bool> & bus_status) const;
 
     virtual void fillSbus(CplxVect & Sbus, const std::vector<int> & id_grid_to_solver, bool ac) const ;